Correcion de login, retorno de tablas parametricas y vistas

+++ src/main/java/co/edu/uniquindio/proyectobases/dto/LoginResponseDto.java
@@ -2,6 +2,6 @@ package co.edu.uniquindio.proyectobases.dto;
 
 public record LoginResponseDto(
     Long idUsuario, 
-    Long idRol, 
+    Integer idRol, 
     String correo
 ) {}
